refactor: refactored typography spacing sizes and tokens
