+++ packages/react/Text/src/components/Skeleton/skeleton.stories.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { ComponentMeta, ComponentStory } from "@storybook/react";
 import { withA11y } from "@storybook/addon-a11y";
 
-import { Skeleton } from "./Skeleton";
+import { Skeleton } from "./Skeletondsad";
 
 export default {
   title: "Skeleton/Text",
